requirements.txt
dill-0.2.7.1.tar.gz
stomp.py-4.1.17.tar.gz